+++ codegen/keyword_generation/generators/class_generator.py
@@ -1,10 +1,31 @@
+import collections
 import os
 import typing
 
 from jinja2 import Environment
 
 import keyword_generation.data_model as data_model
-from keyword_generation.utils import fix_keyword, get_license_header, get_classname
+
+from keyword_generation.handlers.handler_base import KeywordHandler
+from keyword_generation.handlers.external_card import ExternalCardHandler
+from keyword_generation.handlers.series_card import SeriesCardHandler
+from keyword_generation.handlers.shared_field import SharedFieldHandler
+from keyword_generation.handlers.table_card import TableCardHandler
+from keyword_generation.handlers.table_card_group import TableCardGroupHandler
+from keyword_generation.handlers.override_field import OverrideFieldHandler
+from keyword_generation.handlers.rename_property import RenamePropertyHandler
+from keyword_generation.handlers.conditional_card import ConditionalCardHandler
+from keyword_generation.handlers.add_option import AddOptionHandler
+from keyword_generation.handlers.override_subkeyword import OverrideSubkeywordHandler
+from keyword_generation.handlers.skip_card import SkipCardHandler
+from keyword_generation.handlers.insert_card import InsertCardHandler
+from keyword_generation.handlers.replace_card import ReplaceCardHandler
+from keyword_generation.handlers.card_set import CardSetHandler
+from keyword_generation.handlers.reorder_card import ReorderCardHandler
+
+from keyword_generation.utils import fix_keyword, get_license_header, get_classname, handle_single_word_keyword
+
+
 
 def _get_source_keyword(keyword, settings):
     """Get the 'source' keyword to look up in LSPP structs.  Usually
@@ -28,10 +49,244 @@ def _get_jinja_variable(base_variable: typing.Dict) -> typing.Dict:
     )
     return jinja_variable
 
-def _get_base_variable(get_keyword_data, classname: str, keyword: str, keyword_options: typing.Dict) -> typing.Dict:
+def _get_fields(card: typing.Dict) -> typing.List[typing.Dict[str, typing.Any]]:
+    if "duplicate_group" in card:
+        fields = []
+        for sub_card in card["sub_cards"]:
+            fields.extend(sub_card["fields"])
+        return fields
+    return card["fields"]
+
+def _transform_data(data: typing.Dict[str, typing.Any]):
+    """applies the following transformations to data:
+    - lowercase field names (SECID -> secid)
+    - python type mapping (integer->int, real->float)
+    """
+    type_mapping = {"integer": "int", "real": "float", "string": "str", "real-integer": "float"}
+
+    def fix_fieldname(field_name: str) -> str:
+        """returns a python friendly version of field name.
+        For example, nx/ida becomes nx_ida, as becomes as_"""
+        # deal with bad characters
+        for bad_char in ["/", "-", " ", "(", ")", ",", ".", "'", "*", "|", "+"]:
+            field_name = field_name.replace(bad_char, "_")
+        # deal with reserved statements
+        if field_name.lower() in ["global", "as", "int", "lambda", "for"]:
+            field_name = field_name + "_"
+        if field_name[0].isdigit():
+            field_name = "_" + field_name
+        return field_name
+
+    def fix_fieldhelp(field_help: str) -> str:
+        """help is formatted inside a triple quote,"""
+        if field_help.endswith('"'):
+            field_help = field_help + " "
+        # remove any leading whitespace from each line in field_help
+        field_help = "\n".join([l.strip() for l in field_help.split("\n")])
+        return field_help
+
+    def fix_field_string_default(field_default: str) -> str:
+        """string defaults need to be wrapped in quotes"""
+        if field_default == None:
+            return None
+        return f'"{field_default}"'
+
+    def fix_field_int_default(field_default) -> int:
+        """int defaults need to be converted from strings that might be floats"""
+        if field_default == None:
+            return None
+        return int(float(field_default))
+
+    def fix_card(card: typing.Dict) -> None:
+        for field in _get_fields(card):
+            if "used" not in field:
+                field["used"] = True
+            field["type"] = type_mapping[field["type"]]
+            if not field["used"]:
+                field["default"] = None
+                field["help"] = ""
+                field["name"] = "unused"
+                continue
+            if field.get("flag", False):
+                field["type"] = "bool"
+            field_name: str = field["name"]
+            fixed_field_name = fix_fieldname(field_name).lower()
+            if not "property_name" in field:
+                field["property_name"] = fixed_field_name
+            field["name"] = field_name.lower()
+            if not "property_type" in field:
+                field["property_type"] = field["type"]
+            if field["type"] == "str":
+                if "options" in field:
+                    field["options"] = [f'"{option}"' for option in field["options"]]
+                field["default"] = fix_field_string_default(field["default"])
+            elif field["type"] == "int":
+                field["default"] = fix_field_int_default(field["default"])
+            field["help"] = fix_fieldhelp(field["help"])
+
+    index = 0
+    for card in data["cards"]:
+        fix_card(card)
+        card["index"] = index
+        index = index + 1
+
+    card_sets = data.get("card_sets", {})
+    for card_set in card_sets.get("sets", []):
+        for card in card_set.get("source_cards", []):
+            fix_card(card)
+        for option in card_set.get("options", []):
+            [fix_card(card) for card in option["cards"]]
+
+    for option in data.get("options", []):
+        [fix_card(card) for card in option["cards"]]
+
+def _set_keyword_identity(kwd_data:  typing.Dict, keyword_name: str, settings: typing.Dict) -> None:
+    tokens = keyword_name.split("_")
+    kwd_data["keyword"] = tokens[0]
+    kwd_data["subkeyword"] = "_".join(tokens[1:])
+    kwd_data["title"] = handle_single_word_keyword(keyword_name)
+
+
+##########################
+
+# functions which return a copy of keyword data after applying the handling specified by the configuration
+HANDLERS = collections.OrderedDict(
+    {
+        "reorder-card": ReorderCardHandler(),
+        "table-card": TableCardHandler(),
+        "override-field": OverrideFieldHandler(),
+        "replace-card": ReplaceCardHandler(),
+        "insert-card": InsertCardHandler(),
+        "variable-card": SeriesCardHandler(),
+        "add-option": AddOptionHandler(),
+        "card-set": CardSetHandler(),
+        "conditional-card": ConditionalCardHandler(),
+        "rename-property": RenamePropertyHandler(),
+        "skip-card": SkipCardHandler(),
+        "table-card-group": TableCardGroupHandler(),
+        "external-card-implementation": ExternalCardHandler(),
+        "shared-field": SharedFieldHandler(),
+        "override-subkeyword": OverrideSubkeywordHandler(),
+    }
+)
+
+def _do_insertions(kwd_data):
+    # [(a,b,c)] => insert b into c at index a
+    insertion_targets: typing.List[typing.Tuple[int, typing.Dict, typing.List]] = []
+    for insertion in kwd_data["card_insertions"]:
+        insertion: data_model.Insertion = insertion
+        insertion_index = insertion.target_index
+        insertion_name = insertion.target_class
+        insertion_card = insertion.card
+        if insertion_name == "":
+            # insert directly into keyword data
+            container = kwd_data["cards"]
+            for index, card in enumerate(container):
+                card_index = card.get("source_index", card["index"])
+                if card_index == insertion_index:
+                    # we are inserting right before this card, store the index
+                    insertion_targets.append((index, insertion_card, container))
+        else:
+            # insert into another card set
+            card_sets = kwd_data.get("card_sets", {})
+            for card_set in card_sets["sets"]:
+                container = card_set["source_cards"]
+                if card_set["name"] == insertion_name:
+                    found = False
+                    for index, card in enumerate(container):
+                        if card["index"] == insertion_index:
+                            found = True
+                            insertion_targets.append((index, insertion_card, container))
+                    if not found:
+                        insertion_targets.append((len(container), insertion_card, container))
+    for index, item, container in insertion_targets:
+        container.insert(index, item)
+
+def _delete_marked_indices(kwd_data):
+    marked_indices = []
+    for index, card in enumerate(kwd_data["cards"]):
+        if "mark_for_removal" in card:
+            marked_indices.append(index)
+    # removal will affect order if we iterate forwards, so iterate backwards
+    marked_indices.sort(reverse=True)
+    for index in marked_indices:
+        del kwd_data["cards"][index]
+
+    options_list = kwd_data.get("options", [])
+    if len(options_list) > 0:
+        marked_option_indices = []
+        for index, option in enumerate(options_list):
+            if "mark_for_removal" in option:
+                marked_option_indices.append(index)
+        marked_option_indices.sort(reverse=True)
+        for index in marked_option_indices:
+            del options_list[index]
+        if len(options_list) == 0:
+            del kwd_data["options"]
+
+def _add_indices(kwd_data):
+    # handlers might point to cards by a specific index.
+    for index, card in enumerate(kwd_data["cards"]):
+        card["index"] = index
+
+def _prepare_for_insertion(kwd_data):
+    kwd_data["card_insertions"] = []
+
+def _add_option_indices(kwd_data):
+    index = len(kwd_data["cards"])
+    for options in kwd_data.get("options", []):
+        for card in options["cards"]:
+            card["index"] = index
+        index += 1
+
+def _after_handle(kwd_data):
+    _do_insertions(kwd_data)
+    _delete_marked_indices(kwd_data)
+    _add_option_indices(kwd_data)
+    for handler_name, handler in HANDLERS.items():
+        if isinstance(handler, KeywordHandler):
+            handler.post_process(kwd_data)
+
+def _before_handle(kwd_data):
+    _add_indices(kwd_data)
+    _prepare_for_insertion(kwd_data)
+
+
+def _handle_keyword_data(kwd_data, settings):
+    _before_handle(kwd_data)
+    # we have to iterate in the order of the handlers because right now the order still matters
+    # right now this is only true for reorder_card
+    for handler_name, handler in HANDLERS.items():
+        handler_settings = settings.get(handler_name)
+        if handler_settings == None:
+            continue
+        handler.handle(kwd_data, handler_settings)
+    _after_handle(kwd_data)
+##############################
+
+
+
+def _get_keyword_data(keyword_name, keyword, settings):
+    """Gets the keyword data dict from kwdm.  Transforms it
+    based on the generation settings that are passed in, if any,
+    and with default transformations that are needed to produce
+    valid python code.
+    """
+    kwd_data = {"cards": data_model.KWDM_INSTANCE.get_keyword_data_dict(keyword)}
+
+    _set_keyword_identity(kwd_data, keyword_name, settings)
+
+    # transformations based on generation settings
+    _handle_keyword_data(kwd_data, settings)
+
+    # default transformations to a valid format we need for jinja
+    _transform_data(kwd_data)
+    return kwd_data
+
+def _get_base_variable(classname: str, keyword: str, keyword_options: typing.Dict) -> typing.Dict:
     source_keyword = _get_source_keyword(keyword, keyword_options)
     generation_settings = keyword_options.get("generation-options", {})
-    keyword_data = get_keyword_data(keyword, source_keyword, generation_settings)
+    keyword_data = _get_keyword_data(keyword, source_keyword, generation_settings)
     keyword_data["classname"] = classname
     alias = data_model.get_alias(keyword)
     alias_subkeyword = None
@@ -46,11 +301,11 @@ def _get_base_variable(get_keyword_data, classname: str, keyword: str, keyword_o
     }
     return data
 
-def generate_class(get_keyword_data, env: Environment, lib_path: str, item: typing.Dict) -> None:
+def generate_class(env: Environment, lib_path: str, item: typing.Dict) -> None:
     keyword = item["name"]
     fixed_keyword = fix_keyword(keyword)
     classname = item["options"].get("classname", get_classname(fixed_keyword))
-    base_variable = _get_base_variable(get_keyword_data, classname, keyword, item["options"])
+    base_variable = _get_base_variable(classname, keyword, item["options"])
     jinja_variable = _get_jinja_variable(base_variable)
     filename = os.path.join(lib_path, "auto", fixed_keyword.lower() + ".py")
     with open(filename, "w", encoding="utf-8") as f:
